requirements.txt
tk
Pillow
numpy
opencv-python
